+++ server/routes/animalRoutes.js
@@ -5,21 +5,24 @@ const { createAnimal,
     deleteAnimal,
     updateAnimal,
  } = require('../controllers/animalController')
- const {getFilterPets, } = require('../controllers/filteringController') 
+ const {filterAndSortPets, } = require('../controllers/filteringAndSortingController') 
 
  
 const router = express.Router()
 
-router.get('/filter', getFilterPets)
 
 /**
  * Read Only Permission Routes
  */
+// GET filterd/sorted/both animals
+router.get('/filterSort', filterAndSortPets)
+
 // GET all animals
 router.get('/', getAllAnimals)
 
 // GET a single animal
 router.get('/:id', getSingleAnimal)
+
 
 /**
  * Read and Write Permission Routes
